pyproject.toml
[project]
name = "report-flow-studio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-qdrant>=0.1.0",
    "qdrant-client>=1.7.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "typer>=0.21.0",
    "deepagents>=0.3.1",
    "dashscope>=1.25.5",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "python-docx>=1.2.0",
    "sqlmodel>=0.0.31",
    "python-multipart>=0.0.21",
    "pytest>=9.0.2",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
